Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Verified-1
> > 
> > Build Failed 
> > 
> > https://jenkins.onap.org/job/optf-cmso-release-verify/3/ : FAILURE
> > 
> > No problems were identified. If you know why this problem occurred, please add a suitable Cause for it. ( https://jenkins.onap.org/job/optf-cmso-release-verify/3/ )
> > 
> > Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/optf-cmso-release-verify/3
> > 
> > https://jenkins.onap.org/job/optf-cmso-master-verify-java-checkstyle/292/ : SUCCESS
> > 
> > Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/optf-cmso-master-verify-java-checkstyle/292
> 
> I cannot tell why this job is failing.

From the wiki:
Once your self release file was merged and processed by "gerrit-release-verify" and "gerrit-release-merge"...

DO NOT attempt to revert files in releases/  → Even if the release was not needed, we need to keep track on what happened in that repo also the tag needs to be kept in the repo
DO NOT attempt to re-tag the repo with the same version → this will fail as the gerrit-release-merge job already tagged it
DO NOT modify releases files → Once a releases file is merged, it is immediately processed. If the team needs to release a new build number, please bump your versions and generate another docker-stage-release
DO NOT re-use the same stage-release build number for multiple releases files → Once an autorelease package is pushed, it is closed and it cannot be re-released. 
MAKE SURE your project is properly calling the Docker profile in pom while running docker-stage-release. If there are no docker images pushed, the job will fail.
DO NOT cherry-pick release files across branches. Duplicated release files will make the verify and merge job fail as those release were already processed.
Releases from master should be created in master, releases from sub-branches like "el-alto" should be created in "el-alto"


It is not clear why the release verify job is failing. Perhaps because 2.1.0-maven was already processed (unsuccessfully?) on the original merge?
We have never had a successful verify job, yet is appears we have to crate a new release number to move forward? Should it be 2.1.1 ?

Patch-set: 2
Reviewer: Gerrit User 89 <89@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0